+++ cocoa/src/toga_cocoa/app.py
@@ -385,7 +385,7 @@ class App:
             window.content._impl.native.enterFullScreenMode(screen, withOptions=opts)
             # Going full screen causes the window content to be re-homed
             # in a NSFullScreenWindow; teach the new parent window
-            # about it's Toga representations.
+            # about its Toga representations.
             window.content._impl.native.window._impl = window._impl
             window.content._impl.native.window.interface = window