+++ gtk/src/toga_gtk/libs/utils.py
@@ -4,7 +4,7 @@ from . import Gtk
 
 
 def gtk_alignment(alignment):
-    "Convert Toga alignments into arguments compatible with Gtk."
+    """Convert Toga alignments into arguments compatible with Gtk."""
     return {
         LEFT: (0.0, Gtk.Justification.LEFT),
         RIGHT: (1.0, Gtk.Justification.RIGHT),